requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
pyserial
inputs
requests
openai
Pillow
pystray
aiohttp